Massive performance boost to canvas rendering
Turns out, for 5 years we were always scaling the canvas to 5x the size
it needed to be.
I think it was to work around a bug in Safari where it ignores
'canvas-rendering: pixelated'
It still ignores that on old Safari, but whatever. Way faster now! And also renders
on iOS as well.

I also added dynamic scaling to the pixel info box, so it's a more
reasonable size.